+++ integration-test/spark-common/src/test/java/com/tencent/rss/test/SparkClientWithLocalTest.java
@@ -158,7 +158,7 @@ public class SparkClientWithLocalTest extends ShuffleReadWriteBase {
       readClient.readShuffleBlockData();
       fail(EXPECTED_EXCEPTION_MESSAGE);
     } catch (Exception e) {
-      assertTrue(e.getMessage().contains("Failed to read shuffle index"));
+      assertTrue(e.getMessage().contains("Failed to read all replicas for"));
     }
     readClient.close();
   }
